// lib/types.ts
export type Role = "admin" | "commercial";
export type StatutCompte = "en_attente" | "actif" | "desactive" | "rejete";
export type StatutTransfert = "en_attente" | "transferee" | "en_livraison" | "livree" | "annulee";
export type ModePaiement = "especes" | "orange_money" | "mtn_momo" | "wave" | "autre";
export type CategorieProduct = "lunettes_photochromiques" | "lunettes_style" | "montres" | "accessoires";
export type TypeNotification = "info" | "succes" | "alerte" | "erreur";

export interface Profile {
  id: string;
  nom_complet: string;
  telephone: string;
  email?: string;
  ville: string;
  pays: string;
  role: Role;
  statut: StatutCompte;
  photo_url?: string;
  otp_verifie: boolean;
  created_at: string;
  updated_at: string;
}